requirements.txt
asgiref==3.6.0
attrs==22.2.0
Automat==22.10.0
Babel==2.11.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.4
cssselect==1.2.0
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.51.0
docxcompose==1.4.0
docxtpl==0.16.4
filelock==3.8.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.2
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==22.0
parsel==1.7.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
PyJWT==2.6.0
pyOpenSSL==22.1.0
python-docx==0.8.11
python3-openid==3.2.0
pytz==2022.7
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
Scrapy==2.7.1
scrapy-djangoitem==1.1.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.4.0
urllib3==1.26.13
w3lib==2.1.1
zope.interface==5.5.2
